requirements.txt
## Run Script: python run.py -m SGCL -d amazon-beauty
### -m [SGCL, LightGCN, LightGCL]
### -d [ml-100k, amazon-beauty, amazon-toys-games]
### --extract_only: [True, False]

# Kaggle
# 3.11.13 (main, Jun  4 2025, 08:57:29) [GCC 11.4.0]
# PyTorch version: 2.6.0+cu124
# CUDA available: True
# CUDA version (the one PyTorch was built with): 12.4
# GPU name: Tesla P100-PCIE-16GB

# Local
# Python version: 3.11.9
# PyTorch version: 2.5.0+cu124
# CUDA available: True

recbole
torch==2.5.0 --index-url https://download.pytorch.org/whl/cu124
numpy==1.26.0
torch_geometric